Use positional arguments for sync functions

+++ components/widgets/useWidgetSync.ts
@@ -1,4 +1,3 @@
-import { User } from '@supabase/supabase-js';
 import { Dispatch, useEffect, useRef } from 'react';
 import { supabase } from '../supabaseClient';
 import { pageSessionId } from '../syncUtils';
@@ -9,31 +8,39 @@ import widgetSyncService from './widgetSyncService';
 
 // Push the local widget state to the server; this function runs when the
 // widget changes, but also once when there is no widget state on the server
-async function pushLocalWidgetToServer({ state, user }: { state: WidgetState, user: User }) {
-  if (state.isLoading || state.isRemoving) {
+async function pushLocalWidgetToServer(widget: WidgetState) {
+  if (widget.isLoading || widget.isRemoving) {
     return;
   }
-  console.log(`widget push`, state);
+  const user = supabase.auth.user();
+  if (!user) {
+    return;
+  }
+  console.log(`widget push`, widget);
   await supabase
     .from('widgets')
     .upsert([
       {
-        id: state.id,
+        id: widget.id,
         user_id: user.id,
         page_session_id: pageSessionId,
-        raw_data: JSON.stringify(state)
+        raw_data: JSON.stringify(widget)
       }
     ]);
 }
 
 // Delete the widget from the server if it's removed from the local dashboard
-async function deleteLocalWidgetFromServer({ state, user }: { state: WidgetState, user: User }) {
-  console.log(`widget delete`, state);
+async function deleteLocalWidgetFromServer(widget: WidgetState) {
+  const user = supabase.auth.user();
+  if (!user) {
+    return;
+  }
+  console.log(`widget delete`, widget);
   await supabase
     .from('widgets')
     .delete()
     .match({
-      id: state.id,
+      id: widget.id,
       user_id: user.id
     });
 }
@@ -73,10 +80,7 @@ function useWidgetSync(
       if (!user) {
         return;
       }
-      pushLocalWidgetToServer({
-        state: widgetRef.current,
-        user
-      });
+      pushLocalWidgetToServer(widgetRef.current);
     });
     return () => {
       widgetSyncService.offPush(widget.id);
@@ -99,12 +103,8 @@ function useWidgetSync(
   // Delete widget from database when the widget is removed from the local
   // dashboard
   useEffect(() => {
-    const user = supabase.auth.user();
-    if (widget.isRemoving && user) {
-      deleteLocalWidgetFromServer({
-        state: widget,
-        user
-      });
+    if (widget.isRemoving) {
+      deleteLocalWidgetFromServer(widget);
     }
   }, [widget]);
